+++ hata/discord/stage/__init__.py
@@ -5,3 +5,5 @@ __all__ = (
     *preinstanced.__all__,
     *stage.__all__,
 )
+
+# Scheduled events are included within stages, tho they may be moved later.